+++ src/fbc_cv/src/mathematics.cpp
@@ -131,7 +131,7 @@ int64_t av_add_stable(AVRational ts_tb, int64_t ts, AVRational inc_tb, int64_t i
 	int64_t m, d;
 
 	if (inc != 1)
-		inc_tb = av_mul_q(inc_tb, AVRational{ inc, 1 });
+		inc_tb = av_mul_q(inc_tb, AVRational{ static_cast<int>(inc), 1 });
 
 	m = inc_tb.num * (int64_t)ts_tb.den;
 	d = inc_tb.den * (int64_t)ts_tb.num;
